[TASK] Move section "Using sphinx" to reST section
Currently, the section "Using sphinx" only contained a single page: Cross-referencing.
This is confusing to the user, because it splits up one topic (hyperlinks) into 2
sections, one is reST and one in sphinx.

In general it seems beneficial to merge reST and sphinx into one chapter as from the
point of view of someone writing documentation this is easier to handle:

- they may not always know if the directive they are looking for is sphinx or reST
- it is more interesting to them, what the directive can do.