Restructure contribution workflows
In the past, we had the description of the "Edit me on GitHub" workflow on
the start page of the Contribution chapter. This migth result in people
being pushed towards this workflow. And it may be confusing not to
make clear there are 2 different workflows straight up front.

This commit:

- describes the workflows first
- clearly labels and numbers subchapters