[TASK] Add signposting to landing page
Resolves #152

Following community feedback, this commit adds
links to the different sections of this guide on the
homepage along with a brief explanation for each
chapter.

The intent of this commit is to highlight the different
types of documentation that can be created. 
This includes content for extensions, sys extensions,
third party exts and the big one - docs.typo3.org